+++ App.js
@@ -4,10 +4,8 @@ import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
 import { Ionicons } from '@expo/vector-icons';
 import { View, Text, StyleSheet } from 'react-native';
 
-// Importar telas
-// import HomeScreen from './screens/HomeScreen';
-// import ProfileScreen from './screens/ProfileScreen';
-// import SettingsScreen from './screens/SettingsScreen';
+import { HomeScreen } from './lib/resources/pages/home_page';
+import { ProfileScreen } from './lib/resources/pages/profile_page';
 
 // Criar o navegador de abas
 const Tab = createBottomTabNavigator();
@@ -39,44 +37,8 @@ export default function App() {
       >
         <Tab.Screen name="Home" component={HomeScreen} />
         <Tab.Screen name="Profile" component={ProfileScreen} />
-        <Tab.Screen name="Settings" component={SettingsScreen} />
+        {/* <Tab.Screen name="Settings" component={SettingsScreen} /> */}
       </Tab.Navigator>
     </NavigationContainer>
   );
 }
-
-// // Exemplo de tela (HomeScreen.js)
-// import React from 'react';
-// import { View, Text, StyleSheet } from 'react-native';
-
-function HomeScreen() {
-  return (
-    <View style={styles.container}>
-      <Text>Tela Inicial</Text>
-    </View>
-  );
-}
-function ProfileScreen() {
-  return (
-    <View style={styles.container}>
-      <Text>Profile</Text>
-    </View>
-  );
-}
-function SettingsScreen() {
-  return (
-    <View style={styles.container}>
-      <Text>Settings</Text>
-    </View>
-  );
-}
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-});
-
-// export default HomeScreen;
