+++ lib/resources/pages/profile_page.js
@@ -1,0 +1,18 @@
+
+import { View, Text, StyleSheet } from 'react-native';
+
+export  function ProfileScreen() {
+    return (
+        <View style={styles.container}>
+            <Text>Profile</Text>
+        </View>
+    );
+}
+
+const styles = StyleSheet.create({
+    container: {
+        flex: 1,
+        justifyContent: 'center',
+        alignItems: 'center',
+    },
+});
